Fix authorization tags/groups. Fixes #32.
Note that this requires to press the "Update" button for each existing group with auth tags in the admin interface to update the database. Also, the semantics have been changed slightly, so that groups with no read-only tags, but with read/write tags are now readable for the public, but only writable for authorized members.